v1.10.0

+++ nerve/dex/src/main/java/network/nerve/dex/tx/v1/process/TradingOrderProcessor.java
@@ -89,8 +89,10 @@ public class TradingOrderProcessor implements TransactionProcessor {
             TradingOrder order = new TradingOrder();
             order.parse(new NulsByteBuffer(tx.getTxData()));
             TradingOrderPo orderPo = new TradingOrderPo(tx, index, null, order);
-            tradingOrderStorageService.delete(orderPo.getOrderHash());
-            dexManager.removeTradingOrder(orderPo);
+            if(orderPo != null) {
+                tradingOrderStorageService.delete(orderPo.getOrderHash());
+                dexManager.removeTradingOrder(orderPo);
+            }
         } catch (NulsException e) {
             LoggerUtil.dexLog.error("Failure to TradingOrder rollback, hash:" + tx.getHash().toHex());
             LoggerUtil.dexLog.error(e);
